Add files via upload
One of the many applications of the QR decomposition of a matrix is finding its
eigenvalues.
We start by recalling some basic facts from linear algebra. A number  2 C is an
eigenvalue of A 2 Rnn if there is a nonzero vector v 2 Cn such that Av = v. If
A is symmetric then all the eigenvalues of A are real. If Q is an orthogonal matrix
then the eigenvalues of the matrix QTAQ are the same as the eigenvalues of A.